refactor: Delaer와 User를 추상화하는 Player 추상 클래스 생성
